+++ packages/components/input/src/InputMultivalue/InputMultivalue.types.ts
@@ -10,5 +10,6 @@ export interface Props {
   disabled?: boolean;
   onBlur?: () => void;
   onFocus?: () => void;
+  onChange?: (values: React.ReactText[]) => void;
   maxLength?: number;
 }