+++ packages/portal/stories/components/Layout/index.stories.tsx
@@ -1,11 +1,11 @@
 import * as React from 'react';
 import range from 'lodash/range';
-import { boolean, number, select } from '@storybook/addon-knobs';
+import { boolean, number } from '@storybook/addon-knobs';
 
 import Layout from '@synerise/ds-layout';
 import PageHeader from '@synerise/ds-page-header';
 
-const MultilineContent = ({heading, lines = 8}) => (
+const MultilineContent = ({ heading, lines = 8 }) => (
   <React.Fragment>
     <strong>{heading}</strong>
     {range(lines).map(i => (
@@ -16,12 +16,12 @@ const MultilineContent = ({heading, lines = 8}) => (
 
 const stories = {
   default: () => {
+    const [leftOpened, setLeftOpened] = React.useState(false);
+    const [rightOpened, setRightOpened] = React.useState(false);
     const showHeader = boolean('Show header', true);
     const showLeft = boolean('Show left', true);
     const showRight = boolean('Show right', true);
     const showSubheader = boolean('Show subheader', true);
-    const transitionTime = number('Transition time', 100, { range: true, min: 1, max: 10000, step: 1 });
-    const transitionName = select('Transition name', ['default', 'slide'], 'default');
     const leftLines = number('Left lines', 24, { range: true, min: 0, max: 256, step: 1 });
     const rightLines = number('Right lines', 24, { range: true, min: 0, max: 256, step: 1 });
     const contentLines = number('Content lines', 24, { range: true, min: 0, max: 256, step: 1 });
@@ -29,6 +29,8 @@ const stories = {
     return (
       <div style={{ border: '1px dashed gray', height: 440 }}>
         <Layout
+          renderLeftSidebarControls={boolean('Render left sidebar controls', true)}
+          renderRightSidebarControls={boolean('Render right sidebar controls', false)}
           header={
             showHeader && (
               <div>
@@ -37,26 +39,32 @@ const stories = {
             )
           }
           left={
-            showLeft && (
-              <div style={{ padding: 24 }}>
-                <MultilineContent heading="Layout left sidebar" lines={leftLines} />
-              </div>
-            )
+            showLeft && {
+              content: (
+                <div style={{ padding: 24 }}>
+                  <MultilineContent heading="Layout left sidebar (controls visible thus hidden by default)" lines={leftLines} />
+                </div>
+              ),
+              opened: leftOpened,
+              onChange: setLeftOpened,
+            }
           }
           right={
-            showRight && (
-              <div style={{ padding: 24 }}>
-                <MultilineContent heading="Layout right sidebar" lines={rightLines} />
-              </div>
-            )
+            showRight && {
+              content: (
+                <div style={{ padding: 24 }}>
+                  <MultilineContent heading="Layout right sidebar (controls hidden by default)" lines={rightLines} />
+                </div>
+              ),
+              opened: rightOpened,
+              onChange: setRightOpened,
+            }
           }
+          sidebarAnimationDisabled={boolean('Disable sidebar animation', false)}
           subheader={showSubheader && <PageHeader title={'Layout subheader'} bar={'Description'} />}
-          transitionName={transitionName}
-          transitionTime={transitionTime}
         >
           <div style={{ padding: 24 }}>
             <MultilineContent heading="Layout content" lines={contentLines} />
-            
           </div>
         </Layout>
       </div>
@@ -65,7 +73,7 @@ const stories = {
 };
 
 export default {
-name: 'Components/Layout/Layout',
+  name: 'Components/Layout/Layout',
   withoutCenter: true,
   stories,
   Component: Layout,