+++ packages/components/input/CHANGELOG.md
@@ -3,6 +3,33 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.18.6](https://github.com/synerise/synerise-design/compare/@synerise/ds-input@0.18.5...@synerise/ds-input@0.18.6) (2022-02-16)
+
+
+### Bug Fixes
+
+* **input:** InputMultivalue add prop onChange ([16fbfcb](https://github.com/synerise/synerise-design/commit/16fbfcbc1be40471b3a1078d9bf5bbe19b9ce9ed))
+
+
+
+
+
+## [0.18.5](https://github.com/synerise/synerise-design/compare/@synerise/ds-input@0.18.4...@synerise/ds-input@0.18.5) (2022-02-03)
+
+**Note:** Version bump only for package @synerise/ds-input
+
+
+
+
+
+## [0.18.4](https://github.com/synerise/synerise-design/compare/@synerise/ds-input@0.18.3...@synerise/ds-input@0.18.4) (2022-01-17)
+
+**Note:** Version bump only for package @synerise/ds-input
+
+
+
+
+
 ## [0.18.3](https://github.com/synerise/synerise-design/compare/@synerise/ds-input@0.18.2...@synerise/ds-input@0.18.3) (2021-12-22)
 
 **Note:** Version bump only for package @synerise/ds-input