+++ packages/components/page-header/CHANGELOG.md
@@ -3,6 +3,38 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.14.3](https://github.com/synerise/synerise-design/compare/@synerise/ds-page-header@0.14.2...@synerise/ds-page-header@0.14.3) (2022-02-16)
+
+**Note:** Version bump only for package @synerise/ds-page-header
+
+
+
+
+
+## [0.14.2](https://github.com/synerise/synerise-design/compare/@synerise/ds-page-header@0.14.1...@synerise/ds-page-header@0.14.2) (2022-02-03)
+
+**Note:** Version bump only for package @synerise/ds-page-header
+
+
+
+
+
+## [0.14.1](https://github.com/synerise/synerise-design/compare/@synerise/ds-page-header@0.14.0...@synerise/ds-page-header@0.14.1) (2022-01-17)
+
+**Note:** Version bump only for package @synerise/ds-page-header
+
+
+
+
+
+# [0.14.0](https://github.com/synerise/synerise-design/compare/@synerise/ds-page-header@0.13.9...@synerise/ds-page-header@0.14.0) (2022-01-13)
+
+**Note:** Version bump only for package @synerise/ds-page-header
+
+
+
+
+
 ## [0.13.9](https://github.com/synerise/synerise-design/compare/@synerise/ds-page-header@0.13.8...@synerise/ds-page-header@0.13.9) (2022-01-04)
 
 **Note:** Version bump only for package @synerise/ds-page-header