+++ packages/components/popconfirm/CHANGELOG.md
@@ -3,6 +3,33 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.9.2](https://github.com/Synerise/synerise-design/compare/@synerise/ds-popconfirm@0.9.1...@synerise/ds-popconfirm@0.9.2) (2022-02-03)
+
+**Note:** Version bump only for package @synerise/ds-popconfirm
+
+
+
+
+
+## [0.9.1](https://github.com/Synerise/synerise-design/compare/@synerise/ds-popconfirm@0.9.0...@synerise/ds-popconfirm@0.9.1) (2022-01-17)
+
+**Note:** Version bump only for package @synerise/ds-popconfirm
+
+
+
+
+
+# [0.9.0](https://github.com/Synerise/synerise-design/compare/@synerise/ds-popconfirm@0.8.5...@synerise/ds-popconfirm@0.9.0) (2022-01-13)
+
+
+### Bug Fixes
+
+* **popconfirm:** calls onVisibleChange from popconfirm ([ccdee25](https://github.com/Synerise/synerise-design/commit/ccdee25863dc36dbda6620a245ec2916532c293f))
+
+
+
+
+
 ## [0.8.5](https://github.com/Synerise/synerise-design/compare/@synerise/ds-popconfirm@0.8.4...@synerise/ds-popconfirm@0.8.5) (2021-12-29)
 
 **Note:** Version bump only for package @synerise/ds-popconfirm