+++ packages/components/subtle-form/CHANGELOG.md
@@ -3,6 +3,30 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.8.9](https://github.com/Synerise/synerise-design/compare/@synerise/ds-subtle-form@0.8.8...@synerise/ds-subtle-form@0.8.9) (2022-02-16)
+
+**Note:** Version bump only for package @synerise/ds-subtle-form
+
+
+
+
+
+## [0.8.8](https://github.com/Synerise/synerise-design/compare/@synerise/ds-subtle-form@0.8.7...@synerise/ds-subtle-form@0.8.8) (2022-02-03)
+
+**Note:** Version bump only for package @synerise/ds-subtle-form
+
+
+
+
+
+## [0.8.7](https://github.com/Synerise/synerise-design/compare/@synerise/ds-subtle-form@0.8.6...@synerise/ds-subtle-form@0.8.7) (2022-01-17)
+
+**Note:** Version bump only for package @synerise/ds-subtle-form
+
+
+
+
+
 ## [0.8.6](https://github.com/Synerise/synerise-design/compare/@synerise/ds-subtle-form@0.8.5...@synerise/ds-subtle-form@0.8.6) (2021-12-29)
 
 **Note:** Version bump only for package @synerise/ds-subtle-form