+++ packages/components/wizard/CHANGELOG.md
@@ -3,6 +3,46 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.3.4](https://github.com/Synerise/synerise-design/compare/@synerise/ds-wizard@0.3.3...@synerise/ds-wizard@0.3.4) (2022-02-16)
+
+**Note:** Version bump only for package @synerise/ds-wizard
+
+
+
+
+
+## [0.3.3](https://github.com/Synerise/synerise-design/compare/@synerise/ds-wizard@0.3.2...@synerise/ds-wizard@0.3.3) (2022-02-10)
+
+**Note:** Version bump only for package @synerise/ds-wizard
+
+
+
+
+
+## [0.3.2](https://github.com/Synerise/synerise-design/compare/@synerise/ds-wizard@0.3.1...@synerise/ds-wizard@0.3.2) (2022-02-03)
+
+**Note:** Version bump only for package @synerise/ds-wizard
+
+
+
+
+
+## [0.3.1](https://github.com/Synerise/synerise-design/compare/@synerise/ds-wizard@0.3.0...@synerise/ds-wizard@0.3.1) (2022-01-17)
+
+**Note:** Version bump only for package @synerise/ds-wizard
+
+
+
+
+
+# [0.3.0](https://github.com/Synerise/synerise-design/compare/@synerise/ds-wizard@0.2.9...@synerise/ds-wizard@0.3.0) (2022-01-13)
+
+**Note:** Version bump only for package @synerise/ds-wizard
+
+
+
+
+
 ## [0.2.9](https://github.com/Synerise/synerise-design/compare/@synerise/ds-wizard@0.2.8...@synerise/ds-wizard@0.2.9) (2022-01-04)
 
 **Note:** Version bump only for package @synerise/ds-wizard